Move allForks block ops to allForks dir
