+++ packages/beacon-state-transition/src/allForks/block/index.ts
@@ -1,5 +1,9 @@
-export * from "./processBlockHeader";
-export * from "./processEth1Data";
-export * from "./processRandao";
 export * from "./initiateValidatorExit";
 export * from "./isValidIndexedAttestation";
+export * from "./processAttesterSlashing";
+export * from "./processBlockHeader";
+export * from "./processDeposit";
+export * from "./processEth1Data";
+export * from "./processProposerSlashing";
+export * from "./processRandao";
+export * from "./processVoluntaryExit";
