+++ packages/beacon-state-transition/src/altair/block/processDeposit.ts
@@ -2,7 +2,7 @@ import {allForks, altair, phase0} from "@chainsafe/lodestar-types";
 import {ForkName} from "@chainsafe/lodestar-params";
 
 import {CachedBeaconState} from "../../allForks/util";
-import {processDepositAllForks} from "../../phase0/block/processDeposit";
+import {processDeposit as processDepositAllForks} from "../../allForks/block";
 
 export function processDeposit(state: CachedBeaconState<altair.BeaconState>, deposit: phase0.Deposit): void {
   processDepositAllForks(ForkName.altair, state as CachedBeaconState<allForks.BeaconState>, deposit);
