+++ packages/beacon-state-transition/src/phase0/block/processVoluntaryExit.ts
@@ -1,51 +1,11 @@
-import {FAR_FUTURE_EPOCH} from "@chainsafe/lodestar-params";
 import {allForks, phase0} from "@chainsafe/lodestar-types";
-import {isActiveValidator} from "../../util";
 import {CachedBeaconState} from "../../allForks/util";
-import {initiateValidatorExit} from "../../allForks/block";
-import {verifyVoluntaryExitSignature} from "../../allForks/signatureSets";
+import {processVoluntaryExit as processVoluntaryExitAllForks} from "../../allForks/block";
 
 export function processVoluntaryExit(
   state: CachedBeaconState<phase0.BeaconState>,
   signedVoluntaryExit: phase0.SignedVoluntaryExit,
   verifySignature = true
 ): void {
-  assertValidVoluntaryExit(state as CachedBeaconState<allForks.BeaconState>, signedVoluntaryExit, verifySignature);
-  initiateValidatorExit(state as CachedBeaconState<allForks.BeaconState>, signedVoluntaryExit.message.validatorIndex);
-}
-
-export function assertValidVoluntaryExit(
-  state: CachedBeaconState<allForks.BeaconState>,
-  signedVoluntaryExit: phase0.SignedVoluntaryExit,
-  verifySignature = true
-): void {
-  const {config, epochCtx} = state;
-  const voluntaryExit = signedVoluntaryExit.message;
-  const validator = state.validators[voluntaryExit.validatorIndex];
-  const currentEpoch = epochCtx.currentShuffling.epoch;
-  // verify the validator is active
-  if (!isActiveValidator(validator, currentEpoch)) {
-    throw new Error("VoluntaryExit validator is not active");
-  }
-  // verify exit has not been initiated
-  if (validator.exitEpoch !== FAR_FUTURE_EPOCH) {
-    throw new Error("VoluntaryExit validator exit has already been initiated: " + `exitEpoch=${validator.exitEpoch}`);
-  }
-  // exits must specify an epoch when they become valid; they are not valid before then
-  if (!(currentEpoch >= voluntaryExit.epoch)) {
-    throw new Error(
-      "VoluntaryExit epoch is not yet valid: " + `epoch=${voluntaryExit.epoch} currentEpoch=${currentEpoch}`
-    );
-  }
-  // verify the validator had been active long enough
-  if (!(currentEpoch >= validator.activationEpoch + config.SHARD_COMMITTEE_PERIOD)) {
-    throw new Error("VoluntaryExit validator has not been active for long enough");
-  }
-
-  // verify signature
-  if (verifySignature) {
-    if (!verifyVoluntaryExitSignature(state as CachedBeaconState<allForks.BeaconState>, signedVoluntaryExit)) {
-      throw new Error("VoluntaryExit has an invalid signature");
-    }
-  }
+  processVoluntaryExitAllForks(state as CachedBeaconState<allForks.BeaconState>, signedVoluntaryExit, verifySignature);
 }
